+++ config-editor/src-tauri/src/config_editor.rs
@@ -28,10 +28,24 @@ fn stop() {
 #[tauri::command(async)]
 fn start() {
     let ma_exe_path = std::env::current_exe().unwrap();
-    mouse_actions::process_event::process_cmd(vec![
-        ma_exe_path.to_str().unwrap().to_string(),
-        String::from("start"),
-    ])
+    let args = mouse_actions::args::parse();
+    let mut cmd: Vec<String> = Vec::new();
+    cmd.push(ma_exe_path.to_str().unwrap().to_string());
+    // Args::
+    // FIXME : generic forward args
+    if args.no_listen {
+        cmd.push(String::from("--no-listen"));
+    }
+    if args.config_path.is_some() {
+        cmd.push(String::from("--config-path"));
+        cmd.push(args.config_path.unwrap());
+    }
+    if args.log_level.is_some() {
+        cmd.push(String::from("--log-level"));
+        cmd.push(args.log_level.unwrap());
+    }
+    cmd.push(String::from("start"));
+    mouse_actions::process_event::process_cmd(cmd)
 }
 
 #[tauri::command(async)]