+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 * force --no-listen option if Wayland is detected
 * add --log-level option
 
+### Bug Fixes
+
+* forward GUI args to mouse action daemon (the start button)
+
 ## [0.4.1] - 2023-04-21
 
 ### Bug Fixes